core: double cdrom and video devices if both template and instance type used
The problem was that when both template and instance type is used, first
all the devices has been copied from the instance type using the
VmDeviceUtils.copyVmDevices (which creates by default also the cdrom and the
video devices). Than, the disk and and CD device is copied from the template
also using the VmDeviceUtils.copyVmDevices which again by default creates a new
CD and video device. This way they are duplicated.

Fixed by creating a new method VmDeviceUtils.copyDiskDevices() which makes sure
to copy only the disks and do not create any new devices and calling this
method instead of copyVmDevices.

Change-Id: Id6dfbced0b702f3ce0e71283b1e7c4b4cf2604f6
Bug-Url: https://bugzilla.redhat.com/1109880
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>