jsonrpc: Protocol decision based on cluser compatibility version
Engine decides which protocol to use depending on cluster compatibility
version when there was no manual protocol override (UI or REST). When
cluser compatibility verions is set to 3.5 or higher jsonrpc over stomp
is used and for older versions xmlrpc is used.

Next step is to disable user override option when we confirm that
jsonrpc is stable enough.


Change-Id: I63e857be667095f62de7ed5e6a81b25a41e917f4
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>