webadmin: Change trigger for VnicProfileWidget ValueChangeEvent
VnicProfileWidget used to listen on its nameEditor's ValueChangeEvent
in order to fire a ValueChangeEvent on its own. However, when the
nameEditor's value was changed programmatically from its model,
UiCommonEditorVisitor would call an overload of nameEditor.setValue()
that would set the value without firing a ValueChangeEvent.

This patch registers directly to the model, so that the
programmatical change on the model side would trigger a
ValueChangeEvent of VnicProfileWidget.

Change-Id: I330c38ecedb0df6433491bedaeb7afdc2bd11ae5
Bug-Url: https://bugzilla.redhat.com/1114971
Signed-off-by: Lior Vernia <lvernia@redhat.com>