core: change return value of copyImageVDSCommand for better logging
currently the return value of copyImageVDSCommand is not initialized,
and is defaulted with emtpy guid.
this empty guid appears in the engine.log as the result of the command,
which is useless and confusing.

Changed the return value to have the vdsm task id,
which is a better guid to log there.

note that the returnVaule of this command is not in use at all,
only the createInfo, so this affect only the log.

Change-Id: I80d830dcdcdfa38a76b08e5399b2f916349b9b8e
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>