core: drop ConnectSingleAsyncOperation
This patch drops the ConnectSingleAsyncOperation and it's factory with
the simple use of callables.
The ConnectSingleAsyncOperation is more complicated, harder to maintain
and limited than the simple use of callables (for example, results can't
be returned from it at the moment).

There's currently no need for that infrastructure, we can execute the
needed operations from a helper or a helper method.

Change-Id: I825d80da10ef3d22de7cd68a68d854374e6c136c
Signed-off-by: Liron Aravot <laravot@redhat.com>