core: Don't send SetStoragePoolDescription
The storage pool description is never read in VDSM and thus
updating it is a useless (and expensive!) call.
Moreover, this call cannot be trusted - since it's only sent if the
DC's status is UP, there may very well be a situation where the DC's
name in the engine and the pool's description in VDSM do not match.

This patch removes this redundant and harmful call.

Topic: remove-sp-description
Change-Id: I5664c9912d07f9fbb99af315e904270597adbd73
Bug-Url: https://bugzilla.redhat.com/1096215
Signed-off-by: Allon Mureinik <amureini@redhat.com>