core: extend DiskProfile.getAllForStorageDomain to a user query
getAllForStorageDomain is extended to a user query,
which will be consumed by the user portal as well.
In terms of MLA hierarchy, Disk profile is a under Storage Domain (SD),
so all the disk profiles attached to the SD will return if user has permission
to the SD, i.e. for we rely on SD permissions (with no change from current flow).

Change-Id: I3bb75728cd76bba7e39de2314275f5bc87942c6b
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>