core: Added short term history of VM cpu/memory/network usage
The purpose is to have a simple way to remember a short term history of
cpu/mem/network usage. The amount of data remembered is configurable by
UsageHistoryLimit option in vdc_options (by default set to 40 values).

The specific fields in which this data are remembered are in vm_statistics,
namely: memory_usage_history, cpu_usage_history and network_usage_history.

The data are stored as comma separated list.

Change-Id: Id5275c224df83caf5078062e14121847940eac44
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=803251
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>