webadmin: SetupNetwork>edit network(with pencil)>can't edit network
EnumRadioEditor used the global eventBus to fire ValueChangeEvent.
Since UiCommonEditorVisitor subscribes all the editors of type
HasValueChangeHandlers to this event- all the instances of EnumRadioEditor
were subscribed to the same eventBus.
Therefore, each instance pf EnumRadioEditor got the value change events
of the other instances- which caused wrong behaviour and in some cases
ClassCasException.
To fix the issue a separate handlerManger is used for each instance of
EnumRadioEditor.

Note: addValueChangeHandler(..) cannot register to the valueChangeHanlder
of the peer since the peer's 'value' is all the list of enum-items and not
just the selected item as desired.

Change-Id: I83c330b8340ba1cf6b2382877b23d07c25fc1afb
Bug-Url: https://bugzilla.redhat.com/1108014
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>