core: add managed virtio-serial device when reading ovf
virtio-serial devices were written to OVF files before, so when we
import VM/Template and such device exist we'll just add it as managed
device (virtio-serial was added to the list of 'special devices').

It could be that OVF file was created before the unmanaged virtio-serial
device was added. In this case, we'll create a new managed virtio-serial
device for the imported VM/Template.

Change-Id: I08c8dc8514edf5b291017690c654f4a9e455a658
Bug-Url: https://bugzilla.redhat.com/1028387
Signed-off-by: Arik Hadas <ahadas@redhat.com>