core: sync VM config after Live Merge
After a backwards merge, the child of a deleted snapshot will now have a
different image associated with it.  This needs to be reflected in the
VM configuration for the snapshot, else the disks will be marked illegal
in the UI and won't function properly when the snapshot is
previewed/reverted.

Change-Id: Ic37dc1e54b99893bca68088eee80c26e1fbf9f91
Bug-Url: https://bugzilla.redhat.com/1122691
Related-To: https://bugzilla.redhat.com/1083310
Signed-off-by: Greg Padgett <gpadgett@redhat.com>