engine: External network with vlan must be labeled
An external network which its vlan-id is set, must include a label
when it is being added to the external provider.

Since the engine might not be aware of changes made to the provider,
the label values cannot be enforced.

(cherry picked from commit f71b849cd89e819a636a78bae58081c8fa343020)

Change-Id: I811a0d55798b7dd78d37880055c4ac00303b679f
Bug-Url: https://bugzilla.redhat.com/1059249
Signed-off-by: Moti Asayag <masayag@redhat.com>