core: validate osinfo properties files - fail backend on error
validate the osinfo properties files by the order of loading.
on-error, a RuntimeException is thrown, which should fail the backend
from loading.

The parsing itself is implemented using ANTLR parser and lexer
generator and the Osinfo.g4 grammar file.

Osinfo.g4
grammar file that expresses the syntax of the properties file and
except for unique validation needs should be the only place to be
changed when needed.

OsinfoPreferencesParser.java
the tool to parse a property file with and match is against the syntax

  OsinfoPreferencesParser.parse("path/to/file")

Change-Id: Ie33bf9446857e15d13221271120a3ff2b400350c
Bug-Url: https://bugzilla.redhat.com/1056965
Signed-off-by: Roy Golan <rgolan@redhat.com>