restapi: Fix prefix in RSDL
Currently during the build process we generate a RSDL document that
contains "href" attributes including the prefix "/ovirt-engine/api". But
the application is deployed twice, once for "/api" and another for
"/ovirt-engine/api", thus the RSDL document presented by "/api" isn't
valid. To avoid this issue this patch changes the build process so that
it will generate a RSDL document that doesn't contain any prefix. The
correct prefix will be added during runtime by each deployment of the
application.

Change-Id: If5fadb92b6f3a6b06e44729c967289c5917f9a25
Bug-Url: https://bugzilla.redhat.com/1102772
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>