restapi: Faster JAXB element creation
Currently when we need to create a JAXB element we always iterate the
object factory looking for the appropiate method. To improve the
performance of this lookup this patch introduces an index that is
created during startup so that the iteration is done only once.

Change-Id: Ie6f2cc91701ff45a443d3b8b37e8225d7b26c4fa
Bug-Url: https://bugzilla.redhat.com/1250140
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit d078c564be5e1a620850259571052a76052aaed8)