engine: Allow custom properties on NIC without network
Previously a validation failed the Setup Networks command when a NIC
had no network but had custom properties configured, which blocked
removal of networks with custom properties.

Now canDoAction() doesn't block this, but the custom properties are
ignored (similarly to boot protocol) in case no network is configured
on the NIC.

Similar changes have been made to the QoS settings, which would cause
the same problems.

Change-Id: Ibded6b7b7ffe9be960d522bd63cbf9db08b18d6e
Bug-Url: https://bugzilla.redhat.com/1119019
Signed-off-by: Lior Vernia <lvernia@redhat.com>