restapi: Fix access to current user
The code that extracts the current user from the backend stores
VdcQueryReturnValue instead of the DbUser it contains. As a result
several places where the user is needed fail with a NPE, as no object is
found assciated to the DbUser class.

Change-Id: I2dae86a0ab152df2971f55d663cf44f4d239f5e9
Bug-Url: https://bugzilla.redhat.com/1119862
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit c5413a3fc8748d1a260a06a013bf5279799b25e4)