restapi: Different realms for different URLs
Currently we deploy the RESTAPI application twice, once for the /api URL
and another time for the /ovirt-engine/api URL. But both deployments use
the same basic authentication realm. As a result when a browser
authenticates with one URL it will be automatically authenticated with
the other one, as the browser will remember the credentials and send
them automatically if the realm is the same. To avoid this issue this
patch changes the names of the realms, so that they will be different.

Change-Id: I36140bf236b7043f3b813863ce8db635012a11ce
Bug-Url: https://bugzilla.redhat.com/1077441
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 94583560f705e2cbf0dde33c306a9c17add96dd4)