webadmin: Create new VM | Not all selected nics are created on the VM
Opening and closing ghost profile instance type editor without selecting a
value causes the first acceptable value to be displayed at the header text
box (just to be displayed, not really be selected-> the real value is
still null).

It is caused because ListModelTypeAheadListBox->adjustSelectedValue()
calls in the described case-
'setValue(acceptableValues.iterator().next())' without firing valueChange
event. The displayed value is changed, but the model.selectedValue is
still null.

Calling setValue with 'true' for 'fireValueChangeEvent' fixes the issue.

But, as I understand, if the the selected value ('getValue()') was 'null'
and wasn't changed manually to another value.
There is no reason to override it here by the first item in the list.

Change-Id: Iaa8ec33db813f334f07baadac38f27c5544f0791
Bug-Url: https://bugzilla.redhat.com/1115788
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>