engine: Report event for external port removal failure
If an external network provider is inaccessible, the engine
command should not be failed. Instead, the port removal failure
will be logged to the events log for a future reference for the
admin to handle it on the neutron server.

Change-Id: I57d42da98ff1c86f242f53204520fd4e219b3205
Bug-Url: https://bugzilla.redhat.com/1102542
Signed-off-by: Moti Asayag <masayag@redhat.com>
(cherry picked from commit 860309e6d9631acf715b7e44fb458393798ac393)