engine:Change DisplayType semantics
This patch changes meaning of DisplayType. Now this enum determines type
of VM's video device, while graphics type is determined by GraphicsType
(or GraphicsInfo).

The patch also generalizes Run Once handling (instead of useVnc flag in params,
graphics types of graphics cards are listed in RunVmOnceParams).

The REST API part of this patch contains only changes needed for tests to pass.
The actual REST API code was too big and is contained in the follow up patch.

The patch also adds a method to Management Base Commands to derive graphics
that a machine/template will have after adding (the value is not trivial to
find out - by default VM inherits graphics from template/source VM if it's not
customized in parameters). This method is needed for follow up patches:
 - in os info follow up this is needed for validating whether the resulting
   graphics is compliant to os info data,
 - in REST API follow up this method is needed for autoselecting Display Type
   of a VM (user can set Graphics Type via REST API and the Display Type is
   derived according to os info data).

Change-Id: I0907c9086f49f909250d9956356a0251954f1427
Bug-Url: https://bugzilla.redhat.com/1033547
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>