webadmin: Wrong vnics when creating vm from blank template
The 'Blank' template belongs to the 'Default' dc/cluster.
If the 'Blank' template has vnics with profiles, when creating a vm in
different cluster than 'Default' the vnic profiles (which don't exist in
the vm's cluster) should be overridden by 'Empty' profile.

This override happens in EditProfileBehavior->initSelectedProfile, but
since it is called before the registration of ProfileInstanceTypeEditor
to SelectedItemChangedEvent of VnicInstanceType the new selection of
'Empty' profile isn't reflected on VnicInstanceType.NetworkInterface.

This patch fixes it by calling syncSelectedItemWithNetworkInterface(..)
also before the registration and not just from the eventRaised(..).

Change-Id: I60be6b21050c43bb004e56688f8e306b5b943d24
Bug-Url: https://bugzilla.redhat.com/1103058
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>
(cherry picked from commit c57055367478087dbdb3602671f8331b747bd9df)