webadmin: GWT debug mode speedup.
Debug mode would take around 4:30 to get from the login page to
the VM main tab when logging in. This is due to the proxy event
on UiCommonInitEvent in AbstractSubTabPresenter. This event would
fire and it would cause all sub classes of AbstractSubTabPresenter
to initialize and since the code was split, it would have to load
every single sub tab presenter everywhere at once. This took about
500ms-3000ms depending on the presenter. With around 200 sub tab
presenters, it would take a long long time.

This patch removes the proxy event and instead places the handler
initialization in the onBind of the presenter. This should have the
exact same effect as before (the handlers are initalized when the
presenter is activated vs when the application is initalized).

Change-Id: I8a5a038cbf7a6cfe8e462eab2e4f18c910c039cc
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1110788
Signed-off-by: Alexander Wels <awels@redhat.com>