engine: set template id in ctor
When not setting the template id before the execute we have an empty id
for the template.
If we are failing with adding image to the template and the engine is
reset we are deleting the blank template!

When setting the template id right at the beginning prevent this kind of
races.

Bug-Url: https://bugzilla.redhat.com/1105211
Change-Id: I696e572ce87f9fe041a50ca4a88584df385ea353
Signed-off-by: Shahar Havivi <shaharh@redhat.com>