restapi : Bad error message - no link
Incorrect xml posted to restapi results in
a cryptic error message stating

Request syntactically incorrect. See the link
description below for the correct usage:

The link refered to is an internal DetailedLink
entity and should not be shown to the user.

This patch changes the message dropping the
reference to link

Request syntactically incorrect. See the
description below for the correct usage:

Change-Id: Ib38d9c0853bbc487a13057125df8bbb1e9cc77f0
Bug-Url: https://bugzilla.redhat.com/1122591
Signed-off-by: Ravi Nori <rnori@redhat.com>