core: Add memory allocation check to hibernate VM
When hibernating a VM, two volumes are created, one to save memory,
and one for the metadata. When searching for a Storage Domain with
enough space for these volumes, wrong space allocations checks are done.
Added the correct storage allocation checks, replacing existing checks
in HibernateVmCommand.
Added test class to the command - HibernateVmCommandTest, as well as a
test to VmHandlerTest.

Change-Id: I1250e6ea8d67f8026d66cf06589538343d39756a
Bug-Url: https://bugzilla.redhat.com/1124143
Signed-off-by: Vered Volansky <vvolansk@redhat.com>