engine: Remove unused javassisst dependencies
The engine and its test suite do not use Javassisst directly - its is
only used as part of the PowerMock mocking (and is required directly due
to a bug in PowerMock's pom).

Most of PowerMock's usages were removed from the engine a long time ago,
with the exception being the jaxrs module.
This patch removes the leftover javassisst dependencies from all the
other modules that no longer require it.

Change-Id: I360789865b8de899d1a04b0f8a974e95eecc65ca
Signed-off-by: Allon Mureinik <amureini@redhat.com>