restapi: Fix NPE when live migrating a disk
Empty Guid was passed as source domain id to the MoveDisksCommand and
onward. This is fixed and the actual source domain is extracted from
the disk entity when moving/migrating/copying a disk.
The source domain id is then used in the parameters classes.
This was mocked out in the relevant tests.

Change-Id: I570a4c81605b63a860c012e9071e46c34c793635
Bug-Url: https://bugzilla.redhat.com/1124321
Signed-off-by: Vered Volansky <vvolansk@redhat.com>