core: Enable single-entry MAC range in engine-config
This was previously wrongly blocked. It also seems like the utility
class blocking this had another logical issue - in cases where
multiple ranges need to be handled (I don't see that this can happen
at the moment but the code seems to take that into consideration),
true would be returned if at least one of the ranges were valid -
instead of false being returned if at least one of the ranges were
invalid.

Change-Id: I0e439d1b60cc15792ad7e3ce6596f65a78c81b48
Bug-Url: https://bugzilla.redhat.com/1126181
Signed-off-by: Lior Vernia <lvernia@redhat.com>