core: migrating-to is a running state as well
mirgatingTo was a status that wasn't in use.
we started to use the migratingTo status during the migration process.
and many places checking for running() not taking into consideration the
migratingTo, although it should, just as migratingFrom.

in a specific scenario, user was able to remove the vm while it was
migrating to, and caused the vm to become external.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1121089
Change-Id: I39403915fa17439224176fc4e5c065311142485a
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>