core: GetAllDisksByVmId - race condition leads to NPE
The given query is executed to get the vm disks with some
required info (properties from the vm_device, snapshots..etc).
As there are multiple queries on that flow, if some operation is done
while the query runs we might get NPE (and inconsistent results).
This patch prevents NPE when the vm device of a given disk was removed
(which means that it was detached from the vm).

Change-Id: Iffd0f8f695ce88baf2e1fbbe1deedf401a13b246
Bug-Url: https://bugzilla.redhat.com/1130065
Signed-off-by: Liron Aravot <laravot@redhat.com>