aaa: Remove "active" from groups
New optimized and reduced user sync algorithm, is syncing users and
the groups users actually belongs to.
This means that if there is group persisted at the  database which
does not exist at the authz provider,  it will be never reached and
hence not synchronized.
As result, the active/inactive flag in this case is useless as it will
be never changed, nor has any effect.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1120720
Change-Id: Id12a172273228609715c7d09c4d173a83478c82c
Topic: AAA
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>