core: fix fence-status operation
When the fencing policy is null, we check that the host is compatible
with cluster level 3.0, which is wrong. This patch changes that by
not checking cluster level when fencing policy is null. Also
getMinimalSupportedVersion was modified to throw
IllegalArgumentException when null is passed as parameter.

Change-Id: Idecf91da729cbd8744f894c341b7044cc791853d
Bug-Url: https://bugzilla.redhat.com/1130791
Signed-off-by: Oved Ourfali <oourfali@redhat.com>
Signed-off-by: Martin Perina <mperina@redhat.com>
(cherry picked from commit 7039195f1019371a1edbcf3b97a82a098e77c868)