core: fix incorrect value of migrating_to_vds
Since we retried to switch host to maintenance as part of the monitoring
thread but before the updated data from the monitoring was saved, the
migrating_to_vds could have been overridden and set back to its
previous value.

This bug caused NPE in VURTI#handOver method sometimes and sometimes it
produced a VM which is not migrating with some wrong value in its
migrating_to_vds field.

The solution is to retry to switch host to maintenance in the
after-refresh-treatment part of the monitoring.

Change-Id: I423d573ae5251d3058bdfb778a2b486ac5ea0090
Bug-Url: https://bugzilla.redhat.com/1131856
Signed-off-by: Arik Hadas <ahadas@redhat.com>