webadmin: Slightly amend checkbox header logic
Previously the logic only relied on enabled entries to determine
whether the checkbox header should be checked or not, however this
produced inconsistent results when all entries are disabled. Special
attention is now paid to that case.

Also, the logic for determining whether an entry is enabled seemed to
be out-of-date and wasn't functioning properly.

Change-Id: I2e7786d55c0f639c0e2f4af4685d125462724175
Bug-Url: https://bugzilla.redhat.com/1100292
Signed-off-by: Lior Vernia <lvernia@redhat.com>