webadmin: Add UI field for Neutron physical network
The same UI field was being used for both the oVirt label and the
Neutron physical network, and only its interpretation changed
according to context, which was confusing.

As usual, had to change some styling for things to look decent (some
of it is related to unresolved Look & Feel feature issues).

Change-Id: I6b90800cf5948b35283bef03d8e2f95384799c80
Bug-Url: https://bugzilla.redhat.com/1108069
Bug-Url: https://bugzilla.redhat.com/1121546
Signed-off-by: Lior Vernia <lvernia@redhat.com>