engine : Persist Async Task root command in command_entities
Persist async task's root command information
in command entities table.

Modifies the async tasks table and store the root
command id and action type in the async_tasks table
and all other information needs to rebuild the root
command in the command entities table.

The action type is left in the async task table to enable
upgrading from 3.4 to 3.5.1 or later

Change-Id: I7ec5c072c6f3153c9396b0ce0be8e7c59cbefea1
Bug-Url: https://bugzilla.redhat.com/1133041
Signed-off-by: Ravi Nori <rnori@redhat.com>