event map: render optional.
Event map currently must contain 'UNASSIGNED'
for each event that does not have a down event.
This means that the list of events is maintained twice
(AuditLogType and event_map) and occasionally creates bugs[1].
I'm suggesting to make event map optional so that only events
that are actually linked with a down event will be in the table.

Event registration/deletion side (commands) need not be
aware of the existence of event map since only the notifier
uses it.

[1] https://bugzilla.redhat.com/1119285

Change-Id: I93d39249c9b57802948f356b5f0ad470ab7f32b1
Signed-off-by: Mooli Tayer <mtayer@redhat.com>