core: clone vm - if the vm has only sharable disks they are not attached
The problem was that the attachDisks() was called only from endSuccessfully()
which was called only after the async tasks ended. But when there were no
disks to copy, there were no async tasks so this method was not called.

Fixed by moving the logic into the overridden addVmImages() method and
detaching the disks in case of an error during copying the disks.

Change-Id: Ie2a8f4c35c68721e59c0b7e35c9148f148589794
Bug-Url: https://bugzilla.redhat.com/1124340
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>