aaa: Removing groups from users main tab
The groups are removed from the users main tab due
to the following reasons:
1. They are presented at the users sub tab.
2. We are going to present only managed groups
later on in the patchset, as a result we would like
to remove groups column entirely from the users table
and have the group names being calculated based on the
ad_groups table and the group ids that are kept at the
users table.
If we calculate this for every user in the main users
tab we will suffer from a performeance hit.

Change-Id: I53119c28426e678043544105ca69d85ec9c748c1
Topic: AAA
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>