webadmin: changeing vm_type and os_type did reset the dialog
For change of os type:
The problem was that the change of the vm_type in the vm dialog triggered the
InstanceTypeManager.updateFields().

It was not correct since this caused the whole dialog to be re-inited from the
selected instance type/template (depending on the actual selection) and that
should happen only after "big" changes like cluster change which changes
availability of some fields.

Since the vm type only provides some defaults for the fields and not the
availability of the fields, there is no need to go through the instance type
manager - just simply set the fields up.

For the os type:
Similar but a bit more tricky. There are some fields which are actually
provided from instance type. So if the os type changes, in case the selected
instance type is not custom, it should set the "IsChangable" flag and let the
instance type manager to copy the corresponding fields. But not all to avoid
the dialog will be reset again - only the ones related to the os change.

Change-Id: I5859ca02ddcbeb60583a0e79b94f7b9f2647596a
Bug-Url: https://bugzilla.redhat.com/1125344
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>