rest: dont always use vm dynamic values
Currently, when getting vm, some fields always show the dynamic values
(if exists) which contains information that is relevant for the
current/last run, for example, if vm has SPICE as display type, but on
run-once the user choose VNC, VNC will be shown while the vm is running.
another scenario, when user start the vm with SPICE and after that
change to VNC, SPICE is always shown, even if user explicitly request
the next_run info.

in this patch another option is added to the mapper to allow the caller
to decide if he wants the dynamic value or not, and it is used for the
next_run scenario. all other cases were not changed.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1128453
Change-Id: Iae643bd7f887171b389b71d31a36362ba0168709
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>