core: Do not suppress dismissed alerts
No health check alert is issued in ui after first message was issued

This patch fixed a problem when an alert is dismissed by the user.
Current functionality did not addressed dismissed alerts and show for
each host instance one alert from each type.
In the case that the alert was dismissed by the user, it will not show
again.
This patch fix the SP that insert to audit log not to suppress the alert
in case that the alert exists for the host instance but it was dismissed
by the user (delete=true in audit_log table)

Change-Id: Ie831e8ec49b81bcb49f106f2fb377cff4dc4566b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1133611
Signed-off-by: Eli Mesika <emesika@redhat.com>