Checking isProvisioned instead of the redundant entity isDiscovered
isDiscovered was a mistake which checked only the textbox entity. The
entity is filled with value when there are discovered host, so that
check was wrong.

This patch checks if the flow is for discovered or provisioning based on
the isProvisioned entity, which is set when changing the radio buttons.

Change-Id: Iee2f17c2aee18a1a619551a491af56075ae8cdc5
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1133540
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>